FIREFIVE-123614: Support skipping profile level checks for video
[Problem]
Some platforms support a profile / level of video decoder, but do not advertise in video decoder capabilities correctly.

[Solution]
If the app is confident that the profile and level is supported by a platform, and wants to avoid depending on its advertised capabilities, allow app to skip profile and level compatibility check before selecting and creating a video decoder.

cr: https://code.amazon.com/reviews/CR-133652394